refactor: brand, gallery, morelocal 등 위치, 이름 수정(#66)
proviers 폴더에서 providers 폴더로 옮김
routes를 route로 이름 변경
